Connected splash screen to login in screen and new game to player config screen

+++ SpaceTrader/app/src/main/java/com/theLads/spacetrader/entity/Player.java
@@ -91,7 +91,7 @@ public class Player {
 
     public void setEngineerSkill(int skill) {engineerSkill = skill; }
 
-    public void setCredits(int credits) {this.credits = credits;}
+    public void setCredits(double credits) {this.credits = credits;}
 
     public void setShip(Ship ship) {this.ship = ship; }
